Persist module created. Wraps SQLite.
