+++ include/Persist.h
@@ -1,0 +1,6 @@
+#ifndef _REVE_PERSIST_H
+#define _REVE_PERSIST_H
+
+rv_Bool rv_PersistCreateTable(sqlite3 *db, const char *table);
+
+#endif
